improve fetchinfo
